한 곳에 있던 코드 모두 폴더&파일로 구조화
