+++ .gitignore
@@ -22,3 +22,5 @@ client/build/
 # /docs
 docs/
 
+# /backup
+backup/
